# Significantly improve rrule performance on Python 2.x