--- a/dateutil/rrule.py
+++ b/dateutil/rrule.py
@@ -16,7 +16,7 @@ except ImportError:
     from fractions import gcd
 
 from six import advance_iterator, integer_types
-from six.moves import _thread
+from six.moves import _thread, range
 import heapq
 
 # For warning about deprecation of until and count