@pganssle 2016-03-30
pypy is still obscenely slow at running the tests for whatever reason:

This branch:
```
pypy: Ran 904 tests in 94.335s
pypy3: Ran 904 tests in 146.686s
```

Master branch:
```
pypy: Ran 904 tests in 101.686s
pypy3: Ran 904 tests in 156.325s
```

Seems there's no improvement at all. Not sure why pypy is insanely slow.